server/requirements/development.txt
#
# This file is autogenerated by pip-compile with Python 3.12
# by the following command:
#
#    pip-compile requirements/development.in
#
--extra-index-url https://labs.geomatico.es/django-commons/

asgiref==3.8.1
    # via
    #   django
    #   django-cors-headers
autoflake==2.3.1
    # via -r requirements/development.in
autopep8==2.3.1
    # via -r requirements/development.in
cfgv==3.4.0
    # via pre-commit
deprecation==2.1.0
    # via service-commons
distlib==0.3.9
    # via virtualenv
dj-database-url==2.2.0
    # via service-commons
django==4.2.16
    # via
    #   dj-database-url
    #   django-cors-headers
    #   django-debug-toolbar
    #   django-filter
    #   djangorestframework
    #   drf-orjson-renderer
    #   drf-yasg
    #   service-commons
django-cors-headers==4.5.0
    # via service-commons
django-cprofile-middleware==1.0.5
    # via service-commons-test
django-debug-toolbar==4.4.6
    # via service-commons-test
django-environ==0.11.2
    # via service-commons
django-filter==24.3
    # via djangorestframework-gis
djangorestframework==3.15.2
    # via
    #   djangorestframework-gis
    #   drf-orjson-renderer
    #   drf-yasg
    #   service-commons
djangorestframework-gis==1.1
    # via service-commons
drf-orjson-renderer==1.7.3
    # via service-commons
drf-yasg==1.21.7
    # via service-commons
factory-boy==3.3.1
    # via service-commons-test
faker==30.3.0
    # via factory-boy
filelock==3.16.1
    # via virtualenv
flake8==7.1.1
    # via
    #   -r requirements/development.in
    #   flake8-pyproject
flake8-pyproject==1.2.3
    # via -r requirements/development.in
identify==2.6.1
    # via pre-commit
inflection==0.5.1
    # via drf-yasg
iniconfig==2.0.0
    # via pytest
isort==5.13.2
    # via -r requirements/development.in
mccabe==0.7.0
    # via flake8
nodeenv==1.9.1
    # via pre-commit
orjson==3.10.7
    # via drf-orjson-renderer
packaging==24.1
    # via
    #   deprecation
    #   drf-yasg
    #   pytest
platformdirs==4.3.6
    # via virtualenv
pluggy==1.5.0
    # via pytest
pre-commit==4.0.1
    # via -r requirements/development.in
psycopg2==2.9.9
    # via service-commons
pycodestyle==2.12.1
    # via
    #   autopep8
    #   flake8
pyflakes==3.2.0
    # via
    #   autoflake
    #   flake8
pytest==8.3.3
    # via
    #   pytest-django
    #   service-commons-test
pytest-django==4.9.0
    # via service-commons-test
python-dateutil==2.9.0.post0
    # via faker
pytz==2024.2
    # via drf-yasg
pyyaml==6.0.2
    # via
    #   drf-yasg
    #   pre-commit
service-commons==0.2.5
    # via service-commons-test
service-commons-test==0.2.5
    # via -r requirements/development.in
six==1.16.0
    # via python-dateutil
sqlparse==0.5.1
    # via
    #   django
    #   django-debug-toolbar
typing-extensions==4.12.2
    # via
    #   dj-database-url
    #   faker
uritemplate==4.1.1
    # via drf-yasg
virtualenv==20.26.6
    # via pre-commit
